refactor: rename package from htmlwriter to md2tag and update documentation
