+++ build_docs/src/htmlwriter.rst
@@ -1,7 +1,7 @@
-htmlwriter
+md2tag
 ========
 
-.. automodule:: htmlwriter
+.. automodule:: md2tag
    :members:
    :undoc-members:
    :show-inheritance:
