+++ build_docs/src/index.rst
@@ -1,8 +1,8 @@
-htmlwriter
+md2tag
 ==========
 
 
-**htmlwriter** is a Python package for converting Markdown text to HTML. It is a simple package that provides a single class, `MarkdownToHTML`, used to convert Markdown text to HTML.
+**md2tag** is a Python package for converting Markdown text to HTML. It is a simple package that provides a single class, `MarkdownToHTML`, used to convert Markdown text to HTML.
 
 Below, you can find more information on the MarkdownToHTML class source code.
 
@@ -10,14 +10,14 @@ Below, you can find more information on the MarkdownToHTML class source code.
    :maxdepth: 2
    :caption: API Contents:
    
-   htmlwriter
+   md2tag
 
 
 Usage
 -----
 ## Usage
 
-To use htmlwriter you first have your markdown text in the following format:
+To use md2tag you first have your markdown text in the following format:
 
 +-------------------+-------------------------+
 | Markdown Text     | HTML Equivalent         |
@@ -39,7 +39,7 @@ To use htmlwriter you first have your markdown text in the following format:
 
 .. code-block:: python
 
-    from src.htmlwriter import htmlwriter as hw
+    from src.md2tag import md2tag as hw
 
     markdown_text = """
     h. Welcome to MarkdownToHTML
